Accept
Use "Steal UVA test data" to get test data.
Get the answer code on internet named "check.c"
Run my code and the check code to get the difference.
Find the problem that *00 == 0.
Solve this problem by add a difit for calculate "x".
Find the problem that if n not equal 0 but x equal 0.
Solve this problem by compare n and '0' before calculate "x".
And if n equal '0' print 1 and continue.